initial commit changed the bot to queer bot

+++ database/models/match.py
@@ -14,7 +14,7 @@ class MatchStatus(StatusMixin):
 class MatchModel(BaseModel):
     __tablename__ = "matchs"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)  # CHANGED
     sender_id: Mapped[int] = mapped_column(ForeignKey("users.id"))
     receiver_id: Mapped[int] = mapped_column(ForeignKey("users.id"))
     message: Mapped[str] = mapped_column(String(250), nullable=True)
